BugFix: Listen only on localhost
The output claims that the web server is listening on 127.0.0.1, which
would be relatively safe.

However, the code was actually listening on 0.0.0.0. This means that
the markdown preview was being broadcast to the the the local network
(or possibly the internet!) when it was intended to be local, allowing
others to see content that was intended to be private.